Fix remaining timeline float conversion build error

+++ editor/csharp/EditorShell/ViewModels/MainWindowViewModel.cs
@@ -2644,7 +2644,7 @@ public sealed class MainWindowViewModel : INotifyPropertyChanged
                     }
                     else
                     {
-                        TimelineCurrentTime = TimelineDurationSeconds;
+                        TimelineCurrentTime = (float)TimelineDurationSeconds;
                         break;
                     }
                 }
